+++ c_vector.h
@@ -31,17 +31,17 @@ c_vector_h c_vector_init(const struct c_vector_input_init *input_args);
 uint8_t c_vector_free(c_vector_h list);
 
 uint8_t c_vector_insert_in(c_vector_h *list, const void *ele,
-                           const unsigned int index);
+                           const uint8_t index);
 
 const void *c_vector_push(c_vector_h *list, const void *ele);
 
 void *c_vector_find(c_vector_h list, const void *ele);
 
-void *c_vector_get_at_index(c_vector_h list, const unsigned int index);
+void *c_vector_get_at_index(c_vector_h list, const uint8_t index);
 
 uint8_t c_vector_delete_ele(c_vector_h list, const void *ele);
 
-uint8_t c_vector_delete_ele_at_index(c_vector_h list, const unsigned int index);
+uint8_t c_vector_delete_ele_at_index(c_vector_h list, const uint8_t index);
 
 void c_vector_to_string(c_vector_h list);
